requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pyyaml==6.0.1
pydantic==2.5.0
python-dotenv==1.0.0
ollama==0.1.5
